create dialog_states & add TODO
