+++ tg_bot/tasks/__init__.py
@@ -1,4 +1,4 @@
-from . import misc, reset_dilog, personal_settings, image_answer, text_answer
+from . import dialog_states, misc, reset_dilog, personal_settings, image_answer, text_answer
 
 
 def register_all():
