+++ tg_bot/tasks/dialog_states/creators.py
@@ -1,0 +1,20 @@
+from ...base import DISPATCHER
+from .collection import *
+from aiogram.types import Message
+
+
+@DISPATCHER.message_handler(commands=["web_search"])
+async def set_web_search_state(message: Message):
+    await FlagsToTextCreating.web_search.set()
+    await message.reply(
+        text="➡️ Ваше следующее сообщение будет загружено в модель с флагом _Поиск в интернете_",
+        parse_mode="Markdown"
+    )
+
+
+async def set_answer_is_creating_state():
+    await DialogStates.answer_is_creating.set()
+
+
+async def unset_all_states(user_tg_peer_id: int):
+    await DISPATCHER.current_state(chat=user_tg_peer_id, user=user_tg_peer_id).set_state(None)
